+++ apps/studio/src/routes/editor/EditPanel/ChatTab/ChatMessages/AssistantMessage/index.tsx
@@ -1,0 +1,19 @@
+import type { AssistantChatMessageImpl } from '@/lib/editor/engine/chat/message/assistant';
+import { MessageContent } from './MessageContent';
+
+const AssistantMessage = ({ message }: { message: AssistantChatMessageImpl }) => {
+    return (
+        <div className="px-4 py-2 text-small content-start">
+            <div className="flex flex-col text-wrap gap-2">
+                <MessageContent
+                    messageId={message.id}
+                    content={message.content}
+                    applied={message.applied}
+                    isStream={false}
+                />
+            </div>
+        </div>
+    );
+};
+
+export default AssistantMessage;
